+++ cups/clock.c
@@ -1,7 +1,7 @@
 //
 // Monotonic clock API for CUPS.
 //
-// Copyright © 2024-2025 by OpenPrinting.
+// Copyright © 2024-2026 by OpenPrinting.
 //
 // Licensed under Apache License v2.0.  See the file "LICENSE" for more
 // information.
@@ -89,7 +89,7 @@ cupsGetClock(void)
     }
 
     // Convert clock value to seconds...
-    if ((secs = curclock.tv_sec - cups_first_clock.tv_sec + 0.000000001 * (curclock.tv_nsec - cups_first_clock.tv_nsec)) < 0.0)
+    if ((secs = (double)(curclock.tv_sec - cups_first_clock.tv_sec) + 0.000000001 * (double)(curclock.tv_nsec - cups_first_clock.tv_nsec)) < 0.0)
       secs = 0.0;
   }
   else
@@ -105,7 +105,7 @@ cupsGetClock(void)
     }
 
     // Convert time value to seconds...
-    if ((secs = curtime.tv_sec - cups_first_time.tv_sec + 0.000001 * (curtime.tv_usec - cups_first_time.tv_usec)) < 0.0)
+    if ((secs = (double)(curtime.tv_sec - cups_first_time.tv_sec) + 0.000001 * (double)(curtime.tv_usec - cups_first_time.tv_usec)) < 0.0)
       secs = 0.0;
   }
 #endif // _WIN32
