+++ cups/testarray.c
@@ -1,7 +1,7 @@
 //
 // Array test program for CUPS.
 //
-// Copyright © 2020-2024 by OpenPrinting.
+// Copyright © 2020-2026 by OpenPrinting.
 // Copyright © 2007-2014 by Apple Inc.
 // Copyright © 1997-2006 by Easy Software Products.
 //
@@ -480,7 +480,7 @@ get_seconds(void)
 
 
   gettimeofday(&curtime, NULL);
-  return (curtime.tv_sec + 0.000001 * curtime.tv_usec);
+  return ((double)curtime.tv_sec + 0.000001 * (double)curtime.tv_usec);
 }
 #endif // _WIN32
